requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
azure-cognitiveservices-speech==1.34.0
PureCloudPlatformClientV2==180.0.0
python-dotenv==1.0.0
requests==2.31.0
werkzeug==3.0.1
